Remove JMX EP and refactored
