+++ components/apimgt/org.wso2.carbon.apimgt.gateway/src/main/java/org/wso2/carbon/apimgt/gateway/MethodTimeLogger.java
@@ -63,11 +63,11 @@ public class MethodTimeLogger
      *
      * @return true if the property value matches this package name
      */
-    @Pointcut("execution(* *(..)) && !execution(* org.wso2.carbon.apimgt.gateway.CorrelationLogServiceImpl.*(..)) &&" +
-            " if()")
+    @Pointcut("execution(* *(..)) && " +
+            "!execution(* org.wso2.carbon.apimgt.gateway.ConfigurableCorrelationLogService.*(..)) && if()")
     public static boolean pointCutAll() {
         if (!isLogAllSet) {
-            logAllMethods = CorrelationLogServiceImpl.getLogAllMethods();
+            logAllMethods = ConfigurableCorrelationLogService.isLogAllMethods();
             isLogAllSet = true;
         }
         return logAllMethods;
@@ -81,7 +81,7 @@ public class MethodTimeLogger
     @Pointcut("if()")
     public static boolean isConfigEnabled() {
         if (!isSet) {
-            isEnabled = CorrelationLogServiceImpl.isCorrelationLogEnabled();
+            isEnabled = ConfigurableCorrelationLogService.isEnable();
             isSet = true;
         }
         return isEnabled;
