+++ components/apimgt/org.wso2.carbon.apimgt.impl/src/main/java/org/wso2/carbon/apimgt/impl/MethodTimeLogger.java
@@ -61,10 +61,11 @@ public class MethodTimeLogger
      *
      * @return true if the property value matches this package name
      */
-    @Pointcut("execution(* *(..)) && !execution(* org.wso2.carbon.apimgt.impl.CorrelationLogServiceImpl.*(..)) && if()")
+    @Pointcut("execution(* *(..)) && " +
+            "!execution(* org.wso2.carbon.apimgt.impl.ConfigurableCorrelationLogService.*(..)) && if()")
     public static boolean pointCutAll() {
         if (!isLogAllSet) {
-            logAllMethods = CorrelationLogServiceImpl.getLogAllMethods();
+            logAllMethods = ConfigurableCorrelationLogService.isLogAllMethods();
             isLogAllSet = true;
         }
         return logAllMethods;
@@ -78,7 +79,7 @@ public class MethodTimeLogger
     @Pointcut("if()")
     public static boolean isConfigEnabled() {
         if (!isSet) {
-            isEnabled = CorrelationLogServiceImpl.isCorrelationLogEnabled();
+            isEnabled = ConfigurableCorrelationLogService.isEnable();
             isSet = true;
         }
         return isEnabled;
